+++ plugin_info/install.php
@@ -897,6 +897,14 @@
                 // eqLogic: Renaming icons if required
                 $iconsList = array(
                     "IkeaTradfriDimmer" => "Ikea-Tradfri-Dimmer",
+                    'eTRV0100' => 'Danfoss-Ally-Thermostat',
+                    'IkeaTradfriBulbE14WSOpal400lm' => 'Ikea-BulbE14-Globe',
+                    'ProfaluxLigthModule' => 'Profalux-LigthModule',
+                    'Ikea-BulbE14CandleWhite' => 'Ikea-BulbE14-Candle',
+                    'TRADFRIbulbE14Wopch400lm' => 'Ikea-BulbE14-Candle',
+                    'TRADFRIbulbE14WS470lm' => 'Ikea-BulbE14-Candle',
+                    'TRADFRIbulbE27WSopal1000lm' => 'Ikea-BulbE27',
+                    'TRADFRIbulbE27WW806lm' => 'Ikea-BulbE27',
                 );
                 $curIcon = $eqLogic->getConfiguration('ab::icon', '');
                 if (($curIcon != '') && isset($iconsList[$curIcon])) {